+++ plantcv/geospatial/__init__.py
@@ -3,8 +3,14 @@ from plantcv.geospatial.read_geotif import read_geotif
 from plantcv.geospatial.transform_points import transform_points
 from plantcv.geospatial.transform_polygons import transform_polygons
 from plantcv.geospatial.points_to_geojson import points_to_geojson
-from plantcv.geospatial import analyze
 from plantcv.geospatial.points2roi import points2roi_circle
+from plantcv.geospatial.shapes_to_geojson import shapes_to_geojson
+from plantcv.geospatial import analyze
+from plantcv.geospatial import shapes
+from plantcv.geospatial.napari_grid import napari_grid
+from plantcv.geospatial.napari_polygon_grid import napari_polygon_grid
+from plantcv.geospatial.center_grid_rois import center_grid_rois
+from plantcv.geospatial.read_netcdf import read_netcdf
 
 # Auto versioning
 __version__ = version("plantcv-geospatial")
@@ -14,6 +20,13 @@ __all__ = [
     "transform_points",
     "transform_polygons",
     "points_to_geojson",
+    "points2roi_circle",
     "analyze",
-    "points2roi_circle"
+    "points2roi_circle",
+    "shapes_to_geojson",
+    "shapes",
+    "napari_grid",
+    "napari_polygon_grid",
+    "center_grid_rois",
+    "read_netcdf"
 ]