+++ plantcv/geospatial/analyze/__init__.py
@@ -1,5 +1,6 @@
 from plantcv.geospatial.analyze.coverage import coverage
-from plantcv.geospatial.analyze.spectral import spectral
+from plantcv.geospatial.analyze.dsm import height_percentile
+from plantcv.geospatial.analyze.spectral import spectral_index
 
 
-__all__ = ["coverage", "spectral"]
+__all__ = ["coverage", "height_percentile", "spectral_index"]